Adding recovery psw I forgot to put some control, i will add tomorrow

+++ app/src/main/java/com/example/chillinapp/data/stress/FirebaseStressDataDao.kt
@@ -2,8 +2,6 @@ package com.example.chillinapp.data.stress
 
 import android.util.Log
 import com.example.chillinapp.data.ServiceResult
-import com.example.chillinapp.data.account.AccountService
-import com.example.chillinapp.data.account.FirebaseAccountDao
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
 import com.google.firebase.database.FirebaseDatabase
@@ -141,10 +139,10 @@ class FirebaseStressDataDao {
                 if (timestamp != null) {
                     val binterval: Array<Float> = document.get("binterval") as Array<Float>
                     val prediction: Double = (document.get("prediction") as Double)
-                    val stress_level: Float = (document.get("stress_level") as Float)
+                    val stress_level: Double = (document.get("stress_level") as Double)
 
                     // Costruisci l'oggetto StressRawData e aggiungilo alla lista
-                    val stressDerivedData = StressDerivedData(timestamp, binterval, prediction, stress_level)
+                    val stressDerivedData = StressDerivedData(timestamp, binterval, prediction, stress_level.toFloat())
                     derivedDataList.add(stressDerivedData)
                 }
 
@@ -205,6 +203,7 @@ class FirebaseStressDataDao {
         }
 
     }
+
 
     suspend fun avgRawData(sincewhen: Long): ServiceResult<StressRawData?,StressErrorType> {
         val auth = Firebase.auth
